+++ src/contents/settings/protection.tsx
@@ -1,21 +1,59 @@
 import SettingsItem from "@/components/settings-item";
+import DirExclusionsItem from "@/components/settings-item/dir-exclusions";
+import PuaExclusionsItem from "@/components/settings-item/pua-exclusions";
 import { Button } from "@/components/ui/button";
 import { ButtonGroup } from "@/components/ui/button-group";
 import { Label } from "@/components/ui/label";
+import { Skeleton } from "@/components/ui/skeleton";
 import { Switch } from "@/components/ui/switch";
+import useSettings from "@/hooks/use-settings";
+import { DEFAULT_BACKEND_SETTINGS } from "@/lib/settings";
+import { BackendSettings } from "@/lib/types/settings";
 import { cn } from "@/lib/utils";
-import { ArrowDownUp, CheckCircle, CirclePower, Plus, Power, Shield, ShieldOff, Square, Terminal, Trash2, XCircle } from "lucide-react";
-
-const exclusions: {
-     directories: string[],
-     puaCategories: string[]
-} = {
-     directories: ["Thing 1", "Thing 2", "Thing 3"],
-     puaCategories: ["Category 1", "Category 2", "Category 3"],
-}
+import ExclusionsLoader from "@/loaders/components/exclusions";
+import { ArrowDownUp, CheckCircle, CirclePower, Power, Shield, Square, Terminal, XCircle } from "lucide-react";
+import { useTransition, useState, useEffect, useMemo } from "react";
+import { toast } from "sonner";
 
 export default function ProtectionSettings(){
+     const [isFetching, startTransition] = useTransition()
+     const {fetchBackendSettings,setBackendSettings} = useSettings();
+     const [protectionSettings, setProtectionSettings] = useState<BackendSettings["protection"]>(DEFAULT_BACKEND_SETTINGS.protection)
+     useEffect(()=>{
+          startTransition(async()=>{
+               try {
+                    const settings = await fetchBackendSettings("protection")
+                    setProtectionSettings(val=>!settings ? val : settings)
+               } catch (err){
+                    toast.error("Failed to fetch existing protection settings");
+                    console.error(err)
+               }
+          })
+     },[])
+     const updateProtectionSettings = async <K extends keyof BackendSettings["protection"]>(key: K, value: BackendSettings["protection"][K]) => {
+          await setBackendSettings("protection",key,value);
+          setProtectionSettings(prev=>({...prev, [key]: value}))
+     }
+     const handleExclusionAction = async(value: string, key: "dirExclusions" | "puaExclusions", action: "exclude" | "remove") => {
+          let newArr: string[] = []
+          if(key === "dirExclusions") {
+               const {dirExclusions} = protectionSettings
+               const mainArr = !dirExclusions ? DEFAULT_BACKEND_SETTINGS.protection.dirExclusions : dirExclusions
+               newArr = action==="exclude" ? [...mainArr,value] : mainArr.filter(val=>val!==value)
+          } else {
+               const {puaExclusions} = protectionSettings
+               const mainArr = !puaExclusions ? DEFAULT_BACKEND_SETTINGS.protection.puaExclusions : puaExclusions
+               newArr = action==="exclude" ? [...mainArr,value] : mainArr.filter(val=>val!==value)
+          }
+          await updateProtectionSettings(key,newArr);
+          setProtectionSettings(prev=>({
+               ...prev,
+               [key]: newArr
+          }))
+     }
      const isActive = true;
+     const dirExclusions = useMemo(()=>!protectionSettings.dirExclusions ? DEFAULT_BACKEND_SETTINGS.protection.dirExclusions : protectionSettings.dirExclusions,[protectionSettings.dirExclusions]);
+     const puaExclusions = useMemo(()=>!protectionSettings.puaExclusions ? DEFAULT_BACKEND_SETTINGS.protection.puaExclusions : protectionSettings.puaExclusions,[protectionSettings.puaExclusions])
      return (
           <div className="px-1 py-2 space-y-3">
                <SettingsItem
@@ -28,61 +66,17 @@ export default function ProtectionSettings(){
                               <Label>Real-Time Protection</Label>
                               <p className="text-muted-foreground text-sm">Scans the new file once it appeared</p>
                          </div>
-                         <Switch/>
+                         {isFetching ? (
+                              <Skeleton className="w-8 h-[18px]"/>
+                         ) : (
+                              <Switch
+                                   defaultChecked={protectionSettings.realTime || DEFAULT_BACKEND_SETTINGS.protection.realTime}
+                                   checked={protectionSettings.realTime}
+                                   onCheckedChange={checked=>updateProtectionSettings("realTime",checked)}
+                              />
+                         )}
                     </div>
                </SettingsItem>
-               <div className="grid grid-cols-1 md:grid-cols-2 gap-3">
-                    <SettingsItem
-                         Icon={ShieldOff}
-                         title="Exclusions"
-                         description="--exclude-dir"
-                         className="space-y-2.5"
-                    >
-                         <Button variant="outline">
-                              <Plus/>
-                              Add an exclusion
-                         </Button>
-                         {exclusions.directories.length > 0 ? (
-                              <ul className="space-y-2">
-                                   {exclusions.directories.map((exclusion,i)=>(
-                                        <li key={i+1} className="flex justify-between items-center gap-2 pb-1 border-b last:pb-0 last:border-none break-all">
-                                             {exclusion}
-                                             <Button variant="ghost" size="icon-lg" title="Remove">
-                                                  <Trash2/>
-                                             </Button>
-                                        </li>
-                                   ))}
-                              </ul>
-                         ) : (
-                              <p className="text-muted-foreground font-medium text-center mt-1">No Folder Exclusions yet</p>
-                         )}
-                    </SettingsItem>
-                    <SettingsItem
-                         Icon={ShieldOff}
-                         title="PUA Exclusions"
-                         description="--exclude-pua"
-                         className="space-y-2.5"
-                    >
-                         <Button variant="outline">
-                              <Plus/>
-                              Add an exclusion
-                         </Button>
-                         {exclusions.puaCategories.length > 0 ? (
-                              <ul className="space-y-2">
-                                   {exclusions.puaCategories.map((exclusion,i)=>(
-                                        <li key={i+1} className="flex justify-between items-center gap-2 pb-1 border-b last:pb-0 last:border-none break-all">
-                                             {exclusion}
-                                             <Button variant="ghost" size="icon-lg" title="Remove">
-                                                  <Trash2/>
-                                             </Button>
-                                        </li>
-                                   ))}
-                              </ul>
-                         ) : (
-                              <p className="text-muted-foreground font-medium text-center mt-1">No PUA Category Exclusions yet</p>
-                         )}
-                    </SettingsItem>
-               </div>
                <SettingsItem
                     Icon={Shield}
                     title="ClamD (Daemon)"
@@ -115,6 +109,26 @@ export default function ProtectionSettings(){
                          </p>
                     </div>
                </SettingsItem>
+               <div className="grid grid-cols-1 md:grid-cols-2 gap-3">
+                    {isFetching ? (
+                         <ExclusionsLoader items={dirExclusions.length}/>
+                    ) : (
+                         <DirExclusionsItem
+                              data={dirExclusions}
+                              onSubmit={values=>handleExclusionAction(values.path,"dirExclusions","exclude")}
+                              onDelete={path=>handleExclusionAction(path,"dirExclusions","remove")}
+                         />
+                    )}
+                    {isFetching ? (
+                         <ExclusionsLoader items={puaExclusions.length}/>
+                    ) : (
+                         <PuaExclusionsItem
+                              data={puaExclusions}
+                              onSubmit={values=>handleExclusionAction(values.category,"puaExclusions","exclude")}
+                              onDelete={category=>handleExclusionAction(category,"puaExclusions","remove")}
+                         />
+                    )}
+               </div>
           </div>
      )
 }