+++ src/hooks/use-settings.ts
@@ -1,7 +1,9 @@
 import { DEFAULT_SETTINGS } from "@/lib/settings";
-import { ISettings } from "@/lib/types/settings";
+import { store } from "@/lib/settings/store";
+import { BackendSettings, ISettings } from "@/lib/types/settings";
 import { format } from "date-fns";
 import { useState } from "react";
+import { toast } from "sonner";
 
 export default function useSettings(){
      const [settings, setSettings] = useState<ISettings>(()=>{
@@ -13,12 +15,48 @@ export default function useSettings(){
                return DEFAULT_SETTINGS
           }
      });
+     async function getBackendSettings<
+          S extends keyof BackendSettings,
+          K extends keyof BackendSettings[S]
+     >(
+          section: S,
+          key: K
+     ): Promise<BackendSettings[S][K] | undefined> {
+          const data = await store.get<BackendSettings[S]>(section);
+          return data?.[key];
+     }
+     async function fetchBackendSettings<S extends keyof BackendSettings>(section: S): Promise<BackendSettings[S] | undefined> {
+          const data = await store.get<BackendSettings[S]>(section);
+          return data;
+     }
+     async function setBackendSettings<
+          S extends keyof BackendSettings,
+          K extends keyof BackendSettings[S]
+     >(
+          section: S,
+          key: K,
+          value: BackendSettings[S][K]
+     ) {
+          try{
+               const current = (await store.get<BackendSettings[S]>(section)) ?? {};
+               await store.set(section, {
+                    ...current,
+                    [key]: value,
+               });
+          } catch(e){
+               toast.error("Failed to save settings");
+               console.error(e)
+          }
+     }
      const formatDate = (date?: Date) => {
           if(!date) return "Never"
           return format(date,settings.dateFormat)
      }
      return {
           settings,
+          getBackendSettings,
+          setBackendSettings,
+          fetchBackendSettings,
           setSettings: (overrides: Partial<ISettings>) => {
                const newValues: ISettings = {
                     ...settings,