+++ src/lib/types/schema.ts
@@ -1,4 +1,6 @@
 import * as z from "zod"
-import { SchedulerSchema } from "../schemas";
+import { DirExclusionsSchema, PuaExclusionsSchema, SchedulerSchema } from "../schemas";
 
 export type SchedulerType = z.infer<typeof SchedulerSchema>
+export type DirExclusionsType = z.infer<typeof DirExclusionsSchema>
+export type PuaExclusionsType = z.infer<typeof PuaExclusionsSchema>