+++ src/loaders/settings/protection.tsx
@@ -1,3 +1,5 @@
+import ExclusionsLoader from "../components/exclusions";
+
 export default function ProtectionSettingsLoader(){
      return (
           <div className="px-1 py-2 space-y-3">
@@ -10,46 +12,6 @@ export default function ProtectionSettingsLoader(){
                                    <div className="h-3.5 bg-accent rounded-md w-1/4"/>
                               </div>
                               <div className="w-8 h-[18px] bg-accent rounded-md"/>
-                         </div>
-                    </div>
-               </div>
-               <div className="grid grid-cols-1 md:grid-cols-2 gap-3">
-                    <div className="border border-accent animate-pulse flex flex-col gap-4 rounded-xl p-6">
-                         <div className="h-4 bg-accent rounded-md w-1/3"/>
-                         <div className="h-3.5 bg-accent rounded-md w-1/5"/>
-                         <div className="h-9 w-40 bg-accent rounded-md"/>
-                         <div className="space-y-2">
-                              <div className="flex justify-between items-center gap-2 pb-1 border-b last:pb-0 last:border-none">
-                                   <div className="bg-accent rounded-md w-1/2 h-4"/>
-                                   <div className="bg-accent rounded-md size-10"/>
-                              </div>
-                              <div className="flex justify-between items-center gap-2 pb-1 border-b last:pb-0 last:border-none">
-                                   <div className="bg-accent rounded-md w-1/2 h-4"/>
-                                   <div className="bg-accent rounded-md size-10"/>
-                              </div>
-                              <div className="flex justify-between items-center gap-2 pb-1 border-b last:pb-0 last:border-none">
-                                   <div className="bg-accent rounded-md w-1/2 h-4"/>
-                                   <div className="bg-accent rounded-md size-10"/>
-                              </div>
-                         </div>
-                    </div>
-                    <div className="border border-accent animate-pulse flex flex-col gap-4 rounded-xl p-6">
-                         <div className="h-4 bg-accent rounded-md w-1/3"/>
-                         <div className="h-3.5 bg-accent rounded-md w-1/5"/>
-                         <div className="h-9 w-40 bg-accent rounded-md"/>
-                         <div className="space-y-2">
-                              <div className="flex justify-between items-center gap-2 pb-1 border-b last:pb-0 last:border-none">
-                                   <div className="bg-accent rounded-md w-1/2 h-4"/>
-                                   <div className="bg-accent rounded-md size-10"/>
-                              </div>
-                              <div className="flex justify-between items-center gap-2 pb-1 border-b last:pb-0 last:border-none">
-                                   <div className="bg-accent rounded-md w-1/2 h-4"/>
-                                   <div className="bg-accent rounded-md size-10"/>
-                              </div>
-                              <div className="flex justify-between items-center gap-2 pb-1 border-b last:pb-0 last:border-none">
-                                   <div className="bg-accent rounded-md w-1/2 h-4"/>
-                                   <div className="bg-accent rounded-md size-10"/>
-                              </div>
                          </div>
                     </div>
                </div>
@@ -66,6 +28,10 @@ export default function ProtectionSettingsLoader(){
                               <div className="bg-accent rounded-md h-3.5 w-1/2"/>
                          </div>
                     </div>
+               </div>
+               <div className="grid grid-cols-1 md:grid-cols-2 gap-3">
+                    <ExclusionsLoader/>
+                    <ExclusionsLoader/>
                </div>
           </div>
      )